refactor: extract email-field core CSS to styles folder
